requirements.txt
# Core web scraping
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
lxml==4.9.3
feedparser==6.0.10
tqdm==4.66.1
python-dateutil==2.8.2

# OpenAI integration
openai==0.28.1
python-dotenv==1.0.0
tiktoken==0.5.1

# LangChain (core)
langchain==0.0.346

# Pre-compiled vector store (NO COMPILATION NEEDED)
faiss-cpu==1.7.4

# Remove ALL packages that require compilation:
# annoy, hnswlib, chromadb, sentence-transformers, transformers, torch
